+++ demo/src/App.js
@@ -2,8 +2,8 @@ import React, { Component } from 'react'
 import { IntlProvider, addLocaleData } from 'react-intl'
 import ptLocaleData from 'react-intl/locale-data/pt'
 import ptTranslations from '../../src/locales/pt'
-import mockRules from '../../src/rules/BRA'
 import ProfileContainer from '../../src/ProfileContainer'
+import ProfileRules from '../../src/ProfileRules'
 import 'vtex-tachyons'
 
 class App extends Component {
@@ -26,17 +26,37 @@ class App extends Component {
         stateRegistration: null,
         tradeName: null,
       },
+      profileLocale: 'pt-BR',
     }
   }
 
+  toggleLocale = () => {
+    this.setState(prevState => ({
+      profileLocale: prevState.profileLocale === 'pt-BR' ? 'en-US' : 'pt-BR',
+    }))
+  }
+
   render() {
-    const { profile } = this.state
+    const { profile, profileLocale } = this.state
+
+    if (!profile) return null
 
     return (
       <div>
         <h3>ProfileForm demo:</h3>
+        <div className="mb6">
+          <button onClick={this.toggleLocale}>
+            Set rules to {profileLocale === 'pt-BR' ? 'en-US' : 'pt-BR'}
+          </button>
+        </div>
         <IntlProvider locale={'pt-BR'} messages={ptTranslations}>
-          <ProfileContainer profile={profile} rules={mockRules} />
+          <ProfileRules
+            key={profileLocale}
+            locale={profileLocale}
+            fetch={locale => import('../../src/rules/' + locale)}
+          >
+            <ProfileContainer profile={profile} />
+          </ProfileRules>
         </IntlProvider>
       </div>
     )