+++ src/propTypes/RuleShape.js
@@ -3,7 +3,7 @@ import RuleFieldShape from './RuleFieldShape'
 
 export default PropTypes.shape({
   /** The country this rule refers to */
-  country: PropTypes.string.isRequired,
+  country: PropTypes.string,
   /** Configuration for each field of the form */
   fields: PropTypes.arrayOf(RuleFieldShape),
 })